Truth_Sentinel_backend/requirements.txt
fastapi
numpy
uvicorn[standard]
image
tensorflow-cpu
pandas
